streamingccl: Refactor streaming code.
This is a refactoring PR to improve streaming code structure.
In addition, replication stream manager no longer relies on "hooks"
to execute its functionality.

Release Notes: None
