+++ pkg/ccl/BUILD.bazel
@@ -20,7 +20,6 @@ go_library(
         "//pkg/ccl/storageccl",
         "//pkg/ccl/storageccl/engineccl",
         "//pkg/ccl/streamingccl/streamingest",
-        "//pkg/ccl/streamingccl/streamingutils",
         "//pkg/ccl/streamingccl/streamproducer",
         "//pkg/ccl/utilccl",
         "//pkg/ccl/workloadccl",
