+++ pkg/ccl/streamingccl/streamingest/stream_ingestion_job_test.go
@@ -21,12 +21,17 @@ import (
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/streamingccl/streamproducer"
 	"github.com/cockroachdb/cockroach/pkg/jobs"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
+	"github.com/cockroachdb/cockroach/pkg/keys"
+	"github.com/cockroachdb/cockroach/pkg/kv"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/jobutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/skip"
 	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
+	"github.com/cockroachdb/cockroach/pkg/testutils/testcluster"
+	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
@@ -128,4 +133,100 @@ INSERT INTO d.t2 VALUES (2);
 	sourceData := sourceSQL.QueryStr(t, query)
 	destData := hDest.Tenant.SQL.QueryStr(t, query)
 	require.Equal(t, sourceData, destData)
+}
+
+func TestCutoverBuiltin(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+	ctx := context.Background()
+
+	args := base.TestClusterArgs{ServerArgs: base.TestServerArgs{
+		Knobs: base.TestingKnobs{JobsTestingKnobs: jobs.NewTestingKnobsWithShortIntervals()},
+	}}
+	tc := testcluster.StartTestCluster(t, 1, args)
+	defer tc.Stopper().Stop(ctx)
+	registry := tc.Server(0).JobRegistry().(*jobs.Registry)
+	sqlDB := sqlutils.MakeSQLRunner(tc.ServerConn(0))
+	db := sqlDB.DB
+
+	startTimestamp := hlc.Timestamp{WallTime: timeutil.Now().UnixNano()}
+	streamIngestJobRecord := jobs.Record{
+		Description: "test stream ingestion",
+		Username:    security.RootUserName(),
+		Details: jobspb.StreamIngestionDetails{
+			StreamAddress: "randomgen://test",
+			Span:          roachpb.Span{Key: keys.LocalMax, EndKey: keys.LocalMax.Next()},
+			StartTime:     startTimestamp,
+		},
+		Progress: jobspb.StreamIngestionProgress{},
+	}
+	var job *jobs.StartableJob
+	id := registry.MakeJobID()
+	err := tc.Server(0).DB().Txn(ctx, func(ctx context.Context, txn *kv.Txn) (err error) {
+		return registry.CreateStartableJobWithTxn(ctx, &job, id, txn, streamIngestJobRecord)
+	})
+	require.NoError(t, err)
+
+	// Check that sentinel is not set.
+	progress := job.Progress()
+	sp, ok := progress.GetDetails().(*jobspb.Progress_StreamIngest)
+	require.True(t, ok)
+	require.True(t, sp.StreamIngest.CutoverTime.IsEmpty())
+
+	// This should fail since no highwatermark is set on the progress.
+	cutoverTime := timeutil.Now().Round(time.Microsecond)
+	_, err = db.ExecContext(
+		ctx,
+		`SELECT crdb_internal.complete_stream_ingestion_job($1, $2)`,
+		job.ID(), cutoverTime)
+	require.Error(t, err, "cannot cutover to a timestamp")
+
+	var highWater time.Time
+	err = job.Update(ctx, nil, func(_ *kv.Txn, md jobs.JobMetadata, ju *jobs.JobUpdater) error {
+		highWater = timeutil.Now().Round(time.Microsecond)
+		hlcHighWater := hlc.Timestamp{WallTime: highWater.UnixNano()}
+		return jobs.UpdateHighwaterProgressed(hlcHighWater, md, ju)
+	})
+	require.NoError(t, err)
+
+	// This should fail since the highwatermark is less than the cutover time
+	// passed to the builtin.
+	cutoverTime = timeutil.Now().Round(time.Microsecond)
+	_, err = db.ExecContext(
+		ctx,
+		`SELECT crdb_internal.complete_stream_ingestion_job($1, $2)`,
+		job.ID(), cutoverTime)
+	require.Error(t, err, "cannot cutover to a timestamp")
+
+	// Ensure that the builtin runs locally.
+	var explain string
+	err = db.QueryRowContext(ctx,
+		`EXPLAIN SELECT crdb_internal.complete_stream_ingestion_job($1, $2)`, job.ID(),
+		highWater).Scan(&explain)
+	require.NoError(t, err)
+	require.Equal(t, "distribution: local", explain)
+
+	// This should succeed since the highwatermark is equal to the cutover time.
+	var jobID int64
+	err = db.QueryRowContext(
+		ctx,
+		`SELECT crdb_internal.complete_stream_ingestion_job($1, $2)`,
+		job.ID(), highWater).Scan(&jobID)
+	require.NoError(t, err)
+	require.Equal(t, job.ID(), jobspb.JobID(jobID))
+
+	// This should fail since we already have a cutover time set on the job
+	// progress.
+	_, err = db.ExecContext(
+		ctx,
+		`SELECT crdb_internal.complete_stream_ingestion_job($1, $2)`,
+		job.ID(), highWater)
+	require.Error(t, err, "cutover timestamp already set")
+
+	// Check that sentinel is set on the job progress.
+	sj, err := registry.LoadJob(ctx, job.ID())
+	require.NoError(t, err)
+	progress = sj.Progress()
+	sp, ok = progress.GetDetails().(*jobspb.Progress_StreamIngest)
+	require.True(t, ok)
+	require.Equal(t, hlc.Timestamp{WallTime: highWater.UnixNano()}, sp.StreamIngest.CutoverTime)
 }
